% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_data.R
\name{import_data}
\alias{import_data}
\title{Import Data}
\usage{
import_data(file, format)
}
\arguments{
\item{file}{a file}

\item{format}{a forma}
}
\value{
A dataframe from a specified file route and database
}
\description{
Import Data
}
