% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/top_journals.R, R/top_publisher.R
\name{plot_top_journals}
\alias{plot_top_journals}
\title{Top Journal}
\usage{
plot_top_journals(data)

plot_top_journals(data)
}
\arguments{
\item{data}{}
}
\value{
A list of
}
\description{
Top Journal

Top Publisher
}
