% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.R
\name{report}
\alias{report}
\title{report}
\usage{
report(data, n)
}
\arguments{
\item{data}{a publication dataframe}

\item{n}{the number of limit the report to}
}
\value{
a terminal report based on the publication data
}
\description{
report
}
