% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/top_access.R
\name{top_access}
\alias{top_access}
\title{Top Access}
\usage{
top_access(data)
}
\arguments{
\item{data}{A publication dataframe}
}
\value{
A dataframe with the leading access types in the publications
}
\description{
Top Access
}
