% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/top_authors.R
\name{top_authors}
\alias{top_authors}
\title{Top Authors}
\usage{
top_authors(data)
}
\arguments{
\item{data}{A publications dataframe}
}
\value{
A dataframe with the leading authors in the publications
}
\description{
Top Authors
}
