% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/top_city.R
\name{top_city}
\alias{top_city}
\title{Top Authors}
\usage{
top_city(data)
}
\arguments{
\item{data}{A pubblications dataframe}
}
\value{
A data frame with the leading cities in the publications
}
\description{
Top Authors
}
