% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/top_country.R
\name{top_country}
\alias{top_country}
\title{Top Countries}
\usage{
top_country(data)
}
\arguments{
\item{data}{A publications dataframe}
}
\value{
A dataframe with the leading countries in the publications
}
\description{
Top Countries
}
