% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/top_funder.R
\name{top_funder}
\alias{top_funder}
\title{Top Funder}
\usage{
top_funder(data)
}
\arguments{
\item{data}{A publications dataframe}
}
\value{
A dataframe with the leading funders in the publications
}
\description{
Top Funder
}
