% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/top_funder_country.R
\name{top_funder_country}
\alias{top_funder_country}
\title{Top Funder Country}
\usage{
top_funder_country(data)
}
\arguments{
\item{data}{A pubblications dataframe}
}
\value{
A data frame with the leading cities in the publications
}
\description{
Top Funder Country
}
