% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/top_funder_group.R
\name{top_funder_group}
\alias{top_funder_group}
\title{Top Funder Group}
\usage{
top_funder_group(data)
}
\arguments{
\item{data}{A pubblications dataframe}
}
\value{
A data frame with the leading cities in the publications
}
\description{
Top Funder Group
}
