% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/top_journals.R, R/top_publisher.R
\name{top_journals}
\alias{top_journals}
\title{Top Journal}
\usage{
top_journals(data)

top_journals(data)
}
\arguments{
\item{data}{A publication dataframe}
}
\value{
A dataframe with the leading journals in the publications

A dataframe with the leading publishers in the publications
}
\description{
Top Journal

Top Publisher
}
