% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/top_keywords.R
\name{top_keywords}
\alias{top_keywords}
\title{Top Keywords}
\usage{
top_keywords(data)
}
\arguments{
\item{data}{A publications dataframe}
}
\value{
A dataframe with the leading keywords in the publications
}
\description{
Top Keywords
}
