% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/top_org.R
\name{top_org}
\alias{top_org}
\title{Top Research Organizations}
\usage{
top_org(data)
}
\arguments{
\item{data}{}
}
\value{
A dataframe with the leading research organisations in the publications
}
\description{
Top Research Organizations
}
\examples{
file = 
data = import_data(file,"csv")
top_org(data)
}
