% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/top_research_field.R
\name{top_research_field}
\alias{top_research_field}
\title{Top Research Field}
\usage{
top_research_field(data)
}
\arguments{
\item{data}{A publications dataframe}
}
\value{
A dataframe with the leading research fields in the publications
}
\description{
Top Research Field
}
