% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/top_state.R
\name{top_state}
\alias{top_state}
\title{Top Authors}
\usage{
top_state(data)
}
\arguments{
\item{data}{A publications dataframe}
}
\value{
A dataframe with the leading states in the publications
}
\description{
Top Authors
}
