% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/top_type.R
\name{top_type}
\alias{top_type}
\title{Top Publication Type}
\usage{
top_type(data)
}
\arguments{
\item{data}{A publication dataframe}
}
\value{
A dataframe with the leading access types in the publications
}
\description{
Top Publication Type
}
